+++ src/main.py
@@ -3,13 +3,27 @@ import state_manager
 import updater
 import time
 
-interval = 15 * 60 * 60
+version = '1.0.1'
 
 def main():
+    print(f'ClouDNS DyDNS Update Client')
+    print('----------------------------')
+    print(f'Version: {version}')
+    print('2024 Robert Kirchner JR(rjojjr)')
+    print('----------------------------')
+    print()
     if len(sys.argv) == 4 and (sys.argv[1] == '-a' or sys.argv[1] == '--add-hostname'):
         print(f'Adding domain {sys.argv[2]} with update URL {sys.argv[3]}')
         state_manager.add_entry(sys.argv[2], sys.argv[3])
         print(f'Added domain {sys.argv[2]} with update URL {sys.argv[3]}')
+        return
+
+    if len(sys.argv) == 3 and (sys.argv[1] == '-uim' or sys.argv[1] == '--update-interval-minutes'):
+        print(f'Setting update interval minutes to {sys.argv[2]}')
+        config = state_manager.get_config()
+        config['updateIntervalMinutes'] = int(sys.argv[2])
+        state_manager.update_config(config)
+        print(f'Set update interval minutes to {sys.argv[2]}')
         return
 
     print('Starting DyDNS update thread')
@@ -18,6 +32,7 @@ def main():
         state = state_manager.get_state()
         updater.update(state)
         print('Updated ClouDNS domains')
+        interval = state_manager.get_config()['updateIntervalMinutes'] * 60
         time.sleep(interval)