+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Releases
 
+- 1.0.1
+  - Add fix state file not created
+  - Add client is user agnostic in unix
+  - Add ability to configure 'update interval minutes'
+  - Released 11/28/2024
 - 1.0.0
   - Initial Release
   - Released 11/28/2024